Added Method that returns sorting exhibits by beacons distance
the Method ExhibitBusiness.getSortedExhibits(List<Beacon>
sortedBeacons) returns the list of ordered exhibits starting from the
initial HashMap<String, Exhibit> loaded from DB by the SmartMuseumApp
class in the application creation. In the list exhibit activity, at
each beacons ranging detection, we'll call this method to sort the
exhibits depending on the user position, and we'll redraw the list view
of the exhibits.